moved the urdf mapping logic and pose service to hybrid algorithms
